requirements.txt
Django==4.2
djangorestframework==3.14.0
drf_spectacular==0.26.1
python_igraph==0.10.4
